feat: refactor tagline and trends generation logic, introduce new prompts for better content generation
